pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "roomkit"
dynamic = ["version"]
description = "Pure async Python library for multi-channel conversations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Sylvain Boily" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = ["pydantic>=2.9"]

[project.optional-dependencies]
httpx = ["httpx>=0.27"]
pydantic-ai = ["pydantic-ai>=0.1"]
anthropic = ["anthropic>=0.30"]
openai = ["openai>=1.30"]
vllm = ["openai>=1.30"]
gemini = ["google-genai>=1.0.0"]
sherpa-onnx = ["sherpa-onnx>=1.10"]
fastrtc = ["fastrtc", "numpy"]
twilio = ["twilio>=9.0"]
phonenumbers = ["phonenumbers>=8.13"]
pynacl = ["pynacl>=1.5"]
realtime-openai = ["openai>=1.30", "websockets>=13.0"]
realtime-gemini = ["google-genai>=1.0.0"]
websocket = ["websockets>=13.0"]
sse = ["httpx>=0.27", "httpx-sse>=0.4"]
teams = ["botbuilder-core>=4.14"]
whatsapp-personal = ["neonize>=0.3.14"]
postgres = ["asyncpg>=0.29"]
sources = [
    "roomkit[websocket]",
    "roomkit[sse]",
    "roomkit[whatsapp-personal]",
]
providers = [
    "roomkit[httpx]",
    "roomkit[anthropic]",
    "roomkit[openai]",
    "roomkit[vllm]",
    "roomkit[gemini]",
    "roomkit[twilio]",
    "roomkit[teams]",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.27",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "mypy>=1.11",
    "ruff>=0.6",
    "pre-commit>=3.8",
    "roomkit[providers]",
    "roomkit[phonenumbers]",
    "roomkit[pynacl]",
    "roomkit[sources]",
    "roomkit[docs]",
    "roomkit[postgres]",
]
all = [
    "roomkit[providers]",
    "roomkit[pydantic-ai]",
    "roomkit[whatsapp-personal]",
]

[project.urls]
Homepage = "https://roomkit.live"
Repository = "https://github.com/sboily/roomkit"
Issues = "https://github.com/sboily/roomkit/issues"

[tool.hatch.version]
path = "src/roomkit/_version.py"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/docs",
    "/AGENTS.md",
    "/llms.txt",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/roomkit"]
artifacts = [
    "*.md",
    "*.txt",
]

[tool.hatch.build.targets.wheel.force-include]
"AGENTS.md" = "roomkit/AGENTS.md"
"llms.txt" = "roomkit/llms.txt"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "error",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["google", "google.genai", "google.genai.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["sherpa_onnx", "sherpa_onnx.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/roomkit/channels/__init__.py" = ["N802"]  # Factory functions use PascalCase by design

[tool.coverage.run]
source = ["roomkit"]

[tool.coverage.report]
show_missing = true
fail_under = 90
